chamges

+++ database/migrations/2022_07_25_144430_add_receipt_to_payments_table.php
@@ -15,7 +15,7 @@ class AddReceiptToPaymentsTable extends Migration
     {
         Schema::table('payments', function (Blueprint $table) {
             $table->string('receipt')->nullable();
-            $table->integer('amount')->default(0);
+            $table->float('amount',1)->default(0.0);
         });
     }
